Extracted setAttribute and setComposeType to utils. Refactored how to init BaseSchema and StringSchema
